app/pyproject.toml
[tool.poetry]
name = "pipeline-log-management"
version = "0.1.0"
description = ""
authors = ["Luciano Filho <lvgalvaofilho@gmail.com>"]
readme = "README.md"
packages = [{include = "pipeline_log_management"}]

[tool.poetry.dependencies]
python = "3.11.3"
black = "^23.7.0"
loguru = "^0.7.0"
sentry-sdk = "^1.29.2"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
